reorg old tests
