+++ archive_test/camera_triggering/arduino_firmware/arduino_firmware.ino
