+++ archive_test/external_camera_trigger.py
