+++ archive_test/image_processor_multiple.py
