+++ archive_test/magellan_acq.py
