+++ archive_test/regular_tiff_read.py
